add swal confirm to CanelSystemPlanButton

+++ frontend/components/atoms/CanelSystemPlanButton.tsx
@@ -6,22 +6,34 @@ const CanelSystemPlanButton = () => {
   const [cookies] = useCookies(['_square_eight_merchant_session'])
 
   const cancelSystemPlan = () => {
-    axios.delete(`${process.env.BACKEND_URL}/api/internal/accounts/cancel_plan`, {
-      headers: { 
-        'Session-Id': cookies._square_eight_merchant_session
+    swalWithBootstrapButtons.fire({
+      title: '解除します',
+      html: `プランを解除します。<br />よろしいですか？`,
+      icon: 'question',
+      confirmButtonText: '解除する',
+      cancelButtonText: 'キャンセル',
+      showCancelButton: true,
+      showCloseButton: true
+    }).then((result) => {
+      if (result.isConfirmed) {
+        axios.delete(`${process.env.BACKEND_URL}/api/internal/accounts/cancel_plan`, {
+          headers: { 
+            'Session-Id': cookies._square_eight_merchant_session
+          }
+        }).then(response => {
+          swalWithBootstrapButtons.fire({
+            title: '解除しました',
+            icon: 'info'
+          }).then((result) => {
+            location.reload()
+          })
+        }).catch(error => {
+          swalWithBootstrapButtons.fire({
+            title: '解除失敗しました',
+            icon: 'error'
+          })
+        })
       }
-    }).then(response => {
-      swalWithBootstrapButtons.fire({
-        title: '解除しました',
-        icon: 'info'
-      }).then((result) => {
-        location.reload()
-      })
-    }).catch(error => {
-      swalWithBootstrapButtons.fire({
-        title: '解除失敗しました',
-        icon: 'error'
-      })
     })
   }
